+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Changelog
 - About page, available from the system menu ([PR 430](https://github.com/input-output-hk/daedalus/pull/430))
 - Setup default themes for TN and MN and introduce new theme names ([PR 451](https://github.com/input-output-hk/daedalus/pull/451))
 - Transaction fees UX improvements on wallet sent screen ([PR 449](https://github.com/input-output-hk/daedalus/pull/449))
+- Do not block the UI while wallet is being restored/imported ([PR 457](https://github.com/input-output-hk/daedalus/pull/457))
 
 ### Fixes